move KadasGpkgImportDialog class to dedicated file
